% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Plot Population Over Time for a Given Country}
\usage{
CountryPopulation(country_name)
}
\arguments{
\item{country_name}{A character string specifying the name of the country.}
}
\value{
A ggplot object showing the population over time for the specified country.
}
\description{
This function takes a country name as input and returns a line plot of the population over time for that country.
The data is extracted from the WorldPopulation dataset.
}
\examples{
\dontrun{
CountryPopulation("United States")
}

}
